清除项目中的警告

+++ WSFouncDesign/WSFunc/WSPackage/WSBase64/NSData+Base64.m
@@ -20,7 +20,7 @@
     long long inputLength = [inputData length];
     const unsigned char *inputBytes = [inputData bytes];
     
-    long long maxOutputLength = (inputLength / 4 + 1) * 3;
+    NSUInteger maxOutputLength = (NSUInteger)((inputLength / 4 + 1) * 3);
     NSMutableData *outputData = [NSMutableData dataWithLength:maxOutputLength];
     unsigned char *outputBytes = (unsigned char *)[outputData mutableBytes];
 
@@ -49,7 +49,7 @@
     if (accumulator > 2) outputLength++;
     
 
-    outputData.length = outputLength;
+    outputData.length = (NSUInteger)outputLength;
     return outputLength? outputData: nil;
 }
 
@@ -65,7 +65,7 @@
     
     long long maxOutputLength = (inputLength / 3 + 1) * 4;
     maxOutputLength += wrapWidth? (maxOutputLength / wrapWidth) * 2: 0;
-    unsigned char *outputBytes = (unsigned char *)malloc(maxOutputLength);
+    unsigned char *outputBytes = (unsigned char *)malloc((unsigned int)maxOutputLength);
     
     long long i;
     long long outputLength = 0;
@@ -103,8 +103,8 @@
     }
     
 
-    outputBytes = realloc(outputBytes, outputLength);
-    NSString *result = [[NSString alloc] initWithBytesNoCopy:outputBytes length:outputLength encoding:NSASCIIStringEncoding freeWhenDone:YES];
+    outputBytes = realloc(outputBytes, (unsigned long)outputLength);
+    NSString *result = [[NSString alloc] initWithBytesNoCopy:outputBytes length:(NSUInteger)outputLength encoding:NSASCIIStringEncoding freeWhenDone:YES];
 
 #if !__has_feature(objc_arc)
     [result autorelease];
